tehnikum-main/requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-crontab==0.7.1
djangorestframework==3.14.0
filelock==3.9.0
identify==2.5.12
idna==3.4
isort==5.11.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nodeenv==1.7.0
openapi-codec==1.3.2
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==2.21.0
psycopg2==2.9.5
python-decouple==3.6
pytz==2022.7
PyYAML==6.0
requests==2.28.1
simplejson==3.18.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
